Separate dataset from element
